Roll up mass properties

'rollup_mass_props()' rolls up mass properties in a data frame such that the mass properties of each non-leaf vertex element is the aggregation of those of its child elements.

Usage

rollup_mass_props(tree, df, validate_df = validate_mass_props_table, ...)

Arguments

tree

An 'igraph' tree whose vertices are named as the values of the id column of df and whose directed edges point from child id to parent id.

df

A data frame with (at least) these columns: id, mass, Cx, Cy, Cz, Ixx, Iyy, Izz, Ixy, Ixz, Iyz, POIconv, Ipoint.

validate_df

A validator for the tree and table, default validate_mass_props_table()

...

Other parameters passed to rollupTree::rollup()

Value

The updated data frame

Examples

rollup_mass_props(mp_tree, mp_table)[1:100, ]
#>              id         name POIconv       mass          Cx           Cy
#> 1           C.1     System 1       - 635.660192   0.3318148   1.36614630
#> 2         C.1.1    Segment 1       - 372.387816   2.7823660   2.27331548
#> 3         C.1.2    Segment 2       - 263.272375  -3.1343878   0.08299326
#> 4       C.1.1.1    Element 1       -  45.236245   0.6449576   3.40672797
#> 5       C.1.1.2    Element 2       -  83.911615   1.7979454  -0.58033620
#> 6       C.1.1.3    Element 3       - 123.272571   6.4136420   1.63215491
#> 7       C.1.1.4    Element 4       - 119.967384   0.5455573   4.50076013
#> 8       C.1.2.1    Element 5       -  97.128572   2.2245377  -4.18310553
#> 9       C.1.2.2    Element 6       -  46.012547 -23.9286608  -6.13818202
#> 10      C.1.2.3    Element 7       - 120.131257   0.4974119   5.91504913
#> 11    C.1.1.1.1  Subsystem 1       -  24.182758  12.7664845  -7.33521244
#> 12    C.1.1.1.2  Subsystem 2       -  21.053487 -13.2782447  15.74528976
#> 13    C.1.1.2.1  Subsystem 3       -  32.614045  -5.4164053   4.35058525
#> 14    C.1.1.2.2  Subsystem 4       -  18.443508   5.9446314  -2.86799230
#> 15    C.1.1.2.3  Subsystem 5       -  32.854062   6.6317379  -4.19099753
#> 16    C.1.1.3.1  Subsystem 6       -  18.809672 -11.5693537   4.60172214
#> 17    C.1.1.3.2  Subsystem 7       -  21.309572   3.7950516  -3.76444320
#> 18    C.1.1.3.3  Subsystem 8       -  28.107067   3.8710126   2.64280855
#> 19    C.1.1.3.4  Subsystem 9       -  31.461107  16.1782715   8.40798505
#> 20    C.1.1.3.5 Subsystem 10       -  23.585154  13.1261314  -6.10317776
#> 21    C.1.1.4.1 Subsystem 11       -  14.945121  10.1956219   2.09268188
#> 22    C.1.1.4.2 Subsystem 12       -  20.623094 -14.8515213  16.29220017
#> 23    C.1.1.4.3 Subsystem 13       -  27.123877  -2.3767554   2.44839217
#> 24    C.1.1.4.4 Subsystem 14       -  27.512486  12.0096670   3.91406747
#> 25    C.1.1.4.5 Subsystem 15       -  29.762807  -1.5653890  -0.04778306
#> 26    C.1.2.1.1 Subsystem 16       -  22.161742  17.0733845  -8.78939395
#> 27    C.1.2.1.2 Subsystem 17       -  29.542394  -4.0320250 -10.63513308
#> 28    C.1.2.1.3 Subsystem 18       -  19.515486 -12.2676108 -10.14297124
#> 29    C.1.2.1.4 Subsystem 19       -  25.908950   7.5732239  11.60299863
#> 30    C.1.2.2.1 Subsystem 20       -  12.256125 -25.7562192  -0.82404359
#> 31    C.1.2.2.2 Subsystem 21       -  24.446707 -17.8654251  -8.50267866
#> 32    C.1.2.2.3 Subsystem 22       -   9.309714 -37.4443673  -6.92516492
#> 33    C.1.2.3.1 Subsystem 23       -  26.125954   2.3614452   7.99620979
#> 34    C.1.2.3.2 Subsystem 24       -  23.382905  18.2483930  -4.68014506
#> 35    C.1.2.3.3 Subsystem 25       -  35.450403  -3.5173534   6.67001106
#> 36    C.1.2.3.4 Subsystem 26       -  17.672250 -15.1876417   6.45654603
#> 37    C.1.2.3.5 Subsystem 27       -  17.499745  -2.0314207  14.88894819
#> 38  C.1.1.1.1.1   Assembly 1       -  10.071268  15.0192256   8.60615091
#> 39  C.1.1.1.1.2   Assembly 2       -   8.777087  13.1715220 -19.12305251
#> 40  C.1.1.1.1.3   Assembly 3       -   5.334404   7.8469082 -18.03684651
#> 41  C.1.1.1.2.1   Assembly 4       -   8.530216 -27.0839975  -3.93222778
#> 42  C.1.1.1.2.2   Assembly 5       -   3.688640   2.4216967  34.86365323
#> 43  C.1.1.1.2.3   Assembly 6       -   8.834630  -6.5032449  26.76246960
#> 44  C.1.1.2.1.1   Assembly 7       -   3.187103 -20.7902569  -4.91562419
#> 45  C.1.1.2.1.2   Assembly 8       -   7.704628   1.5556086  -1.75343697
#> 46  C.1.1.2.1.3   Assembly 9       -   7.711519  -2.3777753  23.38481702
#> 47  C.1.1.2.1.4  Assembly 10       -   3.549741   2.6316091  31.09013760
#> 48  C.1.1.2.1.5  Assembly 11       -  10.461054 -10.8383769 -11.43556170
#> 49  C.1.1.2.2.1  Assembly 12       -  10.286039  10.1654142  12.63895967
#> 50  C.1.1.2.2.2  Assembly 13       -   2.244389   8.1157966 -57.40848164
#> 51  C.1.1.2.2.3  Assembly 14       -   5.913080  -2.2216845  -9.14137684
#> 52  C.1.1.2.3.1  Assembly 15       -  10.677102  14.8893352 -11.06192763
#> 53  C.1.1.2.3.2  Assembly 16       -   6.552370 -10.5711516 -13.20642056
#> 54  C.1.1.2.3.3  Assembly 17       -   6.344582   1.6405611  16.22198529
#> 55  C.1.1.2.3.4  Assembly 18       -   3.402740  60.7702477  -1.44767560
#> 56  C.1.1.2.3.5  Assembly 19       -   5.877267 -15.1470616  -5.28209204
#> 57  C.1.1.3.1.1  Assembly 20       -   4.293059 -39.5750239 -13.48623549
#> 58  C.1.1.3.1.2  Assembly 21       -   5.084983 -43.6312526 -15.82460462
#> 59  C.1.1.3.1.3  Assembly 22       -   5.945426  -5.0634332  18.67393164
#> 60  C.1.1.3.1.4  Assembly 23       -   3.486204  58.5882609  32.67091018
#> 61  C.1.1.3.2.1  Assembly 24       -   6.550015   9.1370077 -28.74205708
#> 62  C.1.1.3.2.2  Assembly 25       -   5.517889   4.5710202  19.45615975
#> 63  C.1.1.3.2.3  Assembly 26       -   3.595544   2.6384819  29.67219507
#> 64  C.1.1.3.2.4  Assembly 27       -   5.646124  -2.4239229 -18.77436158
#> 65  C.1.1.3.3.1  Assembly 28       -   5.516820  11.9159764   7.58302926
#> 66  C.1.1.3.3.2  Assembly 29       -   5.098215 -11.9097042 -24.95358705
#> 67  C.1.1.3.3.3  Assembly 30       -   9.449215  20.7511141   6.11645689
#> 68  C.1.1.3.3.4  Assembly 31       -   8.042816 -11.4759551  12.66601430
#> 69  C.1.1.3.4.1  Assembly 32       -   6.584066  39.3645423   3.15413640
#> 70  C.1.1.3.4.2  Assembly 33       -   5.508303  20.4291943  24.53406983
#> 71  C.1.1.3.4.3  Assembly 34       -   9.267718  11.8847755  -4.54399695
#> 72  C.1.1.3.4.4  Assembly 35       -   2.768917  -8.3967033  23.54252209
#> 73  C.1.1.3.4.5  Assembly 36       -   7.332103   6.8714831  11.66671923
#> 74  C.1.1.3.5.1  Assembly 37       -  11.593170  17.6646845 -16.62616230
#> 75  C.1.1.3.5.2  Assembly 38       -   7.660895  14.2436968  16.34470506
#> 76  C.1.1.3.5.3  Assembly 39       -   4.331090  -0.9991273 -17.64210148
#> 77  C.1.1.4.1.1  Assembly 40       -   8.329673   4.3469841  19.99215163
#> 78  C.1.1.4.1.2  Assembly 41       -   6.615447  17.5597862 -20.44498292
#> 79  C.1.1.4.2.1  Assembly 42       -   9.221147 -12.9516146  20.11827066
#> 80  C.1.1.4.2.2  Assembly 43       -   4.299463   2.8420127   9.95060449
#> 81  C.1.1.4.2.3  Assembly 44       -   7.102484 -28.0288847  15.16367932
#> 82  C.1.1.4.3.1  Assembly 45       -   4.385272  18.5433795  21.95942075
#> 83  C.1.1.4.3.2  Assembly 46       -   8.071459 -29.7075256  -5.36804768
#> 84  C.1.1.4.3.3  Assembly 47       -   7.148714   9.8136441  11.05020999
#> 85  C.1.1.4.3.4  Assembly 48       -   4.661363  -0.2492873 -18.49730016
#> 86  C.1.1.4.3.5  Assembly 49       -   2.857069   8.7521744   7.23387426
#> 87  C.1.1.4.4.1  Assembly 50       -   6.546347   3.1826705   8.23007957
#> 88  C.1.1.4.4.2  Assembly 51       -   3.715625  34.3777783  11.62050300
#> 89  C.1.1.4.4.3  Assembly 52       -   8.571466  30.4272946  -9.05073687
#> 90  C.1.1.4.4.4  Assembly 53       -   8.679047  -9.0978350  10.16349146
#> 91  C.1.1.4.5.1  Assembly 54       -   4.531672 -27.1616577 -14.73715705
#> 92  C.1.1.4.5.2  Assembly 55       -   8.946426  15.5077442  13.44912404
#> 93  C.1.1.4.5.3  Assembly 56       -   8.407049  -9.8383532  -5.91352500
#> 94  C.1.1.4.5.4  Assembly 57       -   7.877659   2.5984849  -0.66574356
#> 95  C.1.2.1.1.1  Assembly 58       -   3.595711  25.4843378  16.03742296
#> 96  C.1.2.1.1.2  Assembly 59       -   4.942294  26.2242739 -11.55394742
#> 97  C.1.2.1.1.3  Assembly 60       -   4.914863  28.1045404 -28.36321162
#> 98  C.1.2.1.1.4  Assembly 61       -   3.039161 -15.6364131 -24.06411319
#> 99  C.1.2.1.1.5  Assembly 62       -   5.669714  11.7334329   3.03098042
#> 100 C.1.2.1.2.1  Assembly 63       -   4.323378 -40.3180149  20.93528200
#>              Cz         Ixx         Iyy         Izz           Ixy           Ixz
#> 1    -0.6631918 4250838.375 4135081.317 4387869.582 -36939.709460  -5703.634810
#> 2    -2.0480047 2585081.273 2468902.684 2586058.073 -85117.550036 -37973.213859
#> 3     1.2955683 1663292.931 1659055.029 1795672.212  50176.630947  29218.383462
#> 4     2.5048850  320437.392  303359.222  307896.116  11813.853322  12051.383899
#> 5    -7.3675365  604189.223  568712.379  580609.901   2170.787077 -48654.518927
#> 6    -2.6674182  872296.759  815356.633  860561.985 -60933.818719   5139.944339
#> 7     0.5924761  782574.671  774764.842  833089.037 -38926.964140  -7496.647811
#> 8    -2.6720265  575753.163  594699.798  658365.102  -8094.274698  16633.878856
#> 9    -4.3577724  302641.248  282354.062  273741.595   3625.317732   6155.095729
#> 10    6.6687862  770796.067  751263.469  831661.198  60921.950943  12117.654302
#> 11    4.2506333  151242.291  151681.439  140003.007  -9285.981270  11631.081600
#> 12    0.4996586  163041.130  143884.885  154262.950  14334.201792   1519.832868
#> 13   -8.7423181  255264.690  233424.833  236512.404 -11454.292573 -37583.693877
#> 14    1.0825770  134251.664  108878.512  127953.741  -4491.039961  -8773.562504
#> 15  -10.7464962  211601.360  221873.089  212043.694  16207.560493  -1864.143042
#> 16   -0.2197146  126503.158  130821.452  139832.798 -27330.473253  10213.740909
#> 17   -1.9188334  145089.535  116755.199  132187.503 -11581.111520  -1088.805810
#> 18   -8.7607161  201630.697  204023.208  213195.563 -18665.528651  12743.392992
#> 19   -3.9374914  224389.998  198755.692  188721.426 -13754.623083  -3597.508026
#> 20    3.6598721  168849.350  152364.811  172480.750  10479.313700 -12953.599506
#> 21  -19.4235427  112296.219   99628.241  119189.416  -8967.033710 -17431.198538
#> 22   11.4779120  117592.469  131659.484  128036.987  -9033.019787  16001.475802
#> 23   -6.1146799  148774.141  138236.828  178429.755 -10292.287770  -4473.094373
#> 24   10.3500735  171140.859  189129.265  181772.947   2690.691116  -4928.296237
#> 25    0.1932782  216801.699  193574.314  211705.406 -17153.402260    625.542397
#> 26   -3.1579625  135100.032  143906.096  144737.897  11700.687810  14145.752255
#> 27   -1.6034710  178572.190  175405.603  195466.281 -22254.105018   6406.179276
#> 28   -5.7969012   96462.742   96314.603  119650.334   8648.899165   3211.032970
#> 29   -1.1210201  156476.577  167898.906  178778.096  -2639.826917  -6387.783054
#> 30  -10.6351375   87544.807   61464.871   67216.864  10928.233973  -9715.006440
#> 31   -8.5817965  153686.230  155943.843  138841.509  -9668.028348  10857.760966
#> 32   14.9983160   56514.549   57897.933   64554.353   1994.625620   2091.310085
#> 33    7.6261517  155060.550  172731.489  188449.966  -1180.988326  17992.792399
#> 34   -4.6150983  127797.972  130060.541  141189.756   6231.140992  13234.782556
#> 35    0.3587986  236295.355  218763.167  244473.953  29748.102177  -4158.053892
#> 36   25.1766073  133656.236  110567.391  129817.974  14334.394723 -12586.763500
#> 37   14.4091662  102298.616   95136.554  111067.089   6838.233228 -11576.735706
#> 38   -0.3965443   61834.025   78096.997   59572.314  -6527.375078   5395.749384
#> 39    5.6741907   49631.511   48157.705   52108.300   2032.938393   5822.375498
#> 40   10.6821450   34930.904   24789.140   23750.830  -4190.930886    143.800354
#> 41   -4.5443538   76113.709   70474.499   66815.560   8525.678965   9172.832441
#> 42   -3.7039578   16234.087   11931.258   13228.985   -901.438499   -407.744745
#> 43    7.1249663   64299.822   57868.559   65554.321  10793.905985  -6498.123577
#> 44   -0.5215488   25802.120   21426.183   20321.800   7007.142526  -8871.985596
#> 45  -54.2308989   40699.790   28969.192   53216.236   -389.798051  -1967.363656
#> 46   37.2370833   49016.692   53801.316   62102.617 -23667.847151  -8481.410757
#> 47  -30.4156955   26666.256   25282.456   32825.749   9625.752669  -4211.658416
#> 48   -4.2842623   70244.016   67873.034   57809.930  -1798.097754 -16692.198404
#> 49   -0.8658027   67978.979   68562.962   71543.024   5597.424174  -1450.476777
#> 50   14.7472978   11710.319    9451.792    5392.489  -1154.000314   1549.644821
#> 51   -0.7147679   44702.677   29798.364   41047.611  -8224.070232  -8803.942028
#> 52   -1.2035096   68716.185   77494.203   69998.249    682.943523   5380.366191
#> 53    6.1837950   45539.575   37747.562   31778.429   -894.227254  -2205.055746
#> 54  -27.1208730   32001.064   36022.283   33397.266   1286.265320   6640.073682
#> 55  -39.8514468   13767.543    8518.348   19751.283   4102.708394   -805.436120
#> 56  -12.4309214   40413.313   39053.718   37819.265  11438.911688 -16568.649985
#> 57   23.7129294   17914.728   18011.833   14498.468  -2522.745587   1951.342447
#> 58    4.8240934   47107.418   34826.432   42046.974  -7820.613126  -1357.170670
#> 59  -12.0566515   27245.653   31664.948   31844.969  -1038.517625   2406.528967
#> 60  -16.8613960   22398.262   15926.045   17986.742  -3034.105592  -1014.837352
#> 61  -19.4565775   29306.978   36823.491   31638.809  -2419.138856   2246.198531
#> 62  -40.1297937   28681.447   21415.051   31652.751    -54.270761    445.738104
#> 63   61.5224118   15104.515   15808.218   22972.538 -10930.207774  -5434.176365
#> 64   15.3692288   33413.069   16065.064   33156.425   1435.964390      5.324233
#> 65  -11.2910109   33212.756   30248.221   38324.342  -3278.078554   4005.081612
#> 66   -6.5829677   41566.285   42422.412   35744.647 -15428.427820  14420.904167
#> 67  -25.6112022   58658.860   58143.425   73887.573    902.270630  -3914.190103
#> 68   11.3914814   57244.740   60986.944   54086.293    895.060402  -7231.069433
#> 69   12.9315467   47300.076   31742.958   30558.384  -1786.847202  -2448.104881
#> 70  -12.8107747   46172.406   38476.417   26448.545   2420.608542  -3046.753699
#> 71  -14.6217761   47118.477   53262.518   54675.749  -2701.579069   8297.293608
#> 72  -13.8553414   15442.545   17479.214   16522.048  -6762.218355   3692.859012
#> 73    4.8308673   60274.209   47475.963   50518.431  -6085.886006  -7223.677186
#> 74   -2.0005279   77651.725   77646.268   80275.007   8934.508771  -4972.858909
#> 75   37.1182973   42881.466   37620.311   60396.135   5854.943852     25.532832
#> 76  -40.3704964   25251.230   18641.583   24976.286  -3965.704286  -5323.966652
#> 77  -14.9787214   62055.788   64431.582   66092.586  -4340.703949 -14732.151880
#> 78  -25.0201270   43839.614   34181.196   46424.094  -6596.315421  -3188.236242
#> 79    6.0089634   56366.616   73013.899   73292.893   4326.703575   6202.775527
#> 80   -6.6518134   21565.271   17640.749   16774.955  -5840.353150  -1552.799161
#> 81   29.5530054   35334.222   34382.839   35039.624  -7829.141811   8184.825631
#> 82    5.1629471   23808.615   28243.470   25468.843  -3420.554833    255.728694
#> 83  -15.0253938   52783.679   36955.981   50019.347 -10080.945732  -2256.741058
#> 84   -2.6368738   29670.714   24878.994   43039.411   7089.265059   1044.390584
#> 85  -28.3844982   21048.263   25036.480   34783.807   3886.212091    652.602147
#> 86   29.3807428    9464.371    6539.682   10930.743  -3557.958442     42.081181
#> 87    6.2163752   45781.230   52451.069   39868.536   4925.106466   7244.622250
#> 88   34.2572689   10816.878   15469.580    9007.646  -1830.707816  -5213.195232
#> 89   29.3326880   43159.704   47347.169   56508.271 -10980.918893   7969.294805
#> 90  -15.5143314   58130.546   53587.890   65122.796   7776.752360  -4968.301893
#> 91   -3.5262336   22338.766   23560.611   25873.052  -1021.125862  -1676.779659
#> 92   15.3859078   66431.623   66080.346   70565.742  10716.377438  10761.397161
#> 93   -8.3642779   54194.688   38445.743   55862.916  -7242.065427 -11406.215778
#> 94   -5.7882373   67911.593   56173.629   50215.057 -15453.446121   6098.142511
#> 95  -20.3015564   14710.870   20161.217   26662.473   6130.319712  -5113.091626
#> 96  -26.1093416   23748.472   21334.617   21342.264  -3171.608511   1197.094654
#> 97   -9.5517007   24221.324   29050.641   24663.905     -2.231148   7765.307764
#> 98   17.0838864   14448.750   21804.422   19776.953   4805.199488    125.128676
#> 99   22.4133365   43518.472   38061.765   41974.243   4664.189190   5481.732378
#> 100  13.1399591   16230.774   23189.151   24322.460  -2317.667232  -8094.085645
#>               Iyz Ipoint sigma_mass sigma_Cx sigma_Cy sigma_Cz sigma_Ixx
#> 1   -41942.778890  FALSE         NA       NA       NA       NA        NA
#> 2   -24491.138285  FALSE         NA       NA       NA       NA        NA
#> 3   -18581.162253  FALSE         NA       NA       NA       NA        NA
#> 4    -7007.889017  FALSE         NA       NA       NA       NA        NA
#> 5   -26282.536240  FALSE         NA       NA       NA       NA        NA
#> 6     7778.840431  FALSE         NA       NA       NA       NA        NA
#> 7     3282.213215  FALSE         NA       NA       NA       NA        NA
#> 8   -30339.749388  FALSE         NA       NA       NA       NA        NA
#> 9    -5803.382726  FALSE         NA       NA       NA       NA        NA
#> 10   24588.805516  FALSE         NA       NA       NA       NA        NA
#> 11   -3665.048708  FALSE         NA       NA       NA       NA        NA
#> 12   -4317.230079  FALSE         NA       NA       NA       NA        NA
#> 13  -15618.080046  FALSE         NA       NA       NA       NA        NA
#> 14  -12660.695787  FALSE         NA       NA       NA       NA        NA
#> 15    1819.449055  FALSE         NA       NA       NA       NA        NA
#> 16   -8358.576150  FALSE         NA       NA       NA       NA        NA
#> 17   -2664.522232  FALSE         NA       NA       NA       NA        NA
#> 18    6524.161885  FALSE         NA       NA       NA       NA        NA
#> 19    8522.567433  FALSE         NA       NA       NA       NA        NA
#> 20    2207.661280  FALSE         NA       NA       NA       NA        NA
#> 21   -9165.972965  FALSE         NA       NA       NA       NA        NA
#> 22   16261.755436  FALSE         NA       NA       NA       NA        NA
#> 23   -3840.001909  FALSE         NA       NA       NA       NA        NA
#> 24   21945.766469  FALSE         NA       NA       NA       NA        NA
#> 25  -18281.983865  FALSE         NA       NA       NA       NA        NA
#> 26       5.303518  FALSE         NA       NA       NA       NA        NA
#> 27   -9853.100892  FALSE         NA       NA       NA       NA        NA
#> 28    3122.748235  FALSE         NA       NA       NA       NA        NA
#> 29  -22770.952906  FALSE         NA       NA       NA       NA        NA
#> 30   15205.337279  FALSE         NA       NA       NA       NA        NA
#> 31  -16077.685964  FALSE         NA       NA       NA       NA        NA
#> 32   -5237.531391  FALSE         NA       NA       NA       NA        NA
#> 33    6516.436175  FALSE         NA       NA       NA       NA        NA
#> 34   -8520.729962  FALSE         NA       NA       NA       NA        NA
#> 35    9962.012056  FALSE         NA       NA       NA       NA        NA
#> 36   11234.043728  FALSE         NA       NA       NA       NA        NA
#> 37    9468.427630  FALSE         NA       NA       NA       NA        NA
#> 38   -3519.950456  FALSE         NA       NA       NA       NA        NA
#> 39    4559.129833  FALSE         NA       NA       NA       NA        NA
#> 40   -5964.771263  FALSE         NA       NA       NA       NA        NA
#> 41    2156.088143  FALSE         NA       NA       NA       NA        NA
#> 42   -5010.271328  FALSE         NA       NA       NA       NA        NA
#> 43    -267.974948  FALSE         NA       NA       NA       NA        NA
#> 44    5154.870321  FALSE         NA       NA       NA       NA        NA
#> 45   -3477.048727  FALSE         NA       NA       NA       NA        NA
#> 46  -13859.423581  FALSE         NA       NA       NA       NA        NA
#> 47    5127.483013  FALSE         NA       NA       NA       NA        NA
#> 48   -2711.865923  FALSE         NA       NA       NA       NA        NA
#> 49   -2070.208823  FALSE         NA       NA       NA       NA        NA
#> 50    -991.987274  FALSE         NA       NA       NA       NA        NA
#> 51  -11515.303437  FALSE         NA       NA       NA       NA        NA
#> 52  -11916.268826  FALSE         NA       NA       NA       NA        NA
#> 53    4523.028756  FALSE         NA       NA       NA       NA        NA
#> 54   -4774.940657  FALSE         NA       NA       NA       NA        NA
#> 55     295.936924  FALSE         NA       NA       NA       NA        NA
#> 56    9609.927691  FALSE         NA       NA       NA       NA        NA
#> 57   -3248.312482  FALSE         NA       NA       NA       NA        NA
#> 58    -342.568165  FALSE         NA       NA       NA       NA        NA
#> 59   -3073.149111  FALSE         NA       NA       NA       NA        NA
#> 60   -6695.678659  FALSE         NA       NA       NA       NA        NA
#> 61    -131.506377  FALSE         NA       NA       NA       NA        NA
#> 62    3742.217686  FALSE         NA       NA       NA       NA        NA
#> 63   -5155.417997  FALSE         NA       NA       NA       NA        NA
#> 64    3015.467649  FALSE         NA       NA       NA       NA        NA
#> 65   -2656.345652  FALSE         NA       NA       NA       NA        NA
#> 66   12337.712023  FALSE         NA       NA       NA       NA        NA
#> 67   11840.405547  FALSE         NA       NA       NA       NA        NA
#> 68  -14301.486339  FALSE         NA       NA       NA       NA        NA
#> 69   12811.505498  FALSE         NA       NA       NA       NA        NA
#> 70    -847.222299  FALSE         NA       NA       NA       NA        NA
#> 71   -1158.885711  FALSE         NA       NA       NA       NA        NA
#> 72    1007.568218  FALSE         NA       NA       NA       NA        NA
#> 73   -3585.739955  FALSE         NA       NA       NA       NA        NA
#> 74   10782.268288  FALSE         NA       NA       NA       NA        NA
#> 75    6936.382566  FALSE         NA       NA       NA       NA        NA
#> 76   -6866.109439  FALSE         NA       NA       NA       NA        NA
#> 77   -1595.625647  FALSE         NA       NA       NA       NA        NA
#> 78   -6073.206514  FALSE         NA       NA       NA       NA        NA
#> 79    1578.772095  FALSE         NA       NA       NA       NA        NA
#> 80    5635.594394  FALSE         NA       NA       NA       NA        NA
#> 81    9203.878220  FALSE         NA       NA       NA       NA        NA
#> 82   -3168.856714  FALSE         NA       NA       NA       NA        NA
#> 83    -940.782920  FALSE         NA       NA       NA       NA        NA
#> 84    2868.150821  FALSE         NA       NA       NA       NA        NA
#> 85    2829.401355  FALSE         NA       NA       NA       NA        NA
#> 86   -1027.319079  FALSE         NA       NA       NA       NA        NA
#> 87    1654.211542  FALSE         NA       NA       NA       NA        NA
#> 88     -55.162112  FALSE         NA       NA       NA       NA        NA
#> 89    2941.045792  FALSE         NA       NA       NA       NA        NA
#> 90   14461.092328  FALSE         NA       NA       NA       NA        NA
#> 91    2797.328702  FALSE         NA       NA       NA       NA        NA
#> 92  -20852.971541  FALSE         NA       NA       NA       NA        NA
#> 93   -4273.136308  FALSE         NA       NA       NA       NA        NA
#> 94    6580.012003  FALSE         NA       NA       NA       NA        NA
#> 95    1467.005707  FALSE         NA       NA       NA       NA        NA
#> 96    3473.574464  FALSE         NA       NA       NA       NA        NA
#> 97   -3901.398540  FALSE         NA       NA       NA       NA        NA
#> 98   -1665.416137  FALSE         NA       NA       NA       NA        NA
#> 99     803.880061  FALSE         NA       NA       NA       NA        NA
#> 100  -2867.232937  FALSE         NA       NA       NA       NA        NA
#>     sigma_Iyy sigma_Izz sigma_Ixy sigma_Ixz sigma_Iyz
#> 1          NA        NA        NA        NA        NA
#> 2          NA        NA        NA        NA        NA
#> 3          NA        NA        NA        NA        NA
#> 4          NA        NA        NA        NA        NA
#> 5          NA        NA        NA        NA        NA
#> 6          NA        NA        NA        NA        NA
#> 7          NA        NA        NA        NA        NA
#> 8          NA        NA        NA        NA        NA
#> 9          NA        NA        NA        NA        NA
#> 10         NA        NA        NA        NA        NA
#> 11         NA        NA        NA        NA        NA
#> 12         NA        NA        NA        NA        NA
#> 13         NA        NA        NA        NA        NA
#> 14         NA        NA        NA        NA        NA
#> 15         NA        NA        NA        NA        NA
#> 16         NA        NA        NA        NA        NA
#> 17         NA        NA        NA        NA        NA
#> 18         NA        NA        NA        NA        NA
#> 19         NA        NA        NA        NA        NA
#> 20         NA        NA        NA        NA        NA
#> 21         NA        NA        NA        NA        NA
#> 22         NA        NA        NA        NA        NA
#> 23         NA        NA        NA        NA        NA
#> 24         NA        NA        NA        NA        NA
#> 25         NA        NA        NA        NA        NA
#> 26         NA        NA        NA        NA        NA
#> 27         NA        NA        NA        NA        NA
#> 28         NA        NA        NA        NA        NA
#> 29         NA        NA        NA        NA        NA
#> 30         NA        NA        NA        NA        NA
#> 31         NA        NA        NA        NA        NA
#> 32         NA        NA        NA        NA        NA
#> 33         NA        NA        NA        NA        NA
#> 34         NA        NA        NA        NA        NA
#> 35         NA        NA        NA        NA        NA
#> 36         NA        NA        NA        NA        NA
#> 37         NA        NA        NA        NA        NA
#> 38         NA        NA        NA        NA        NA
#> 39         NA        NA        NA        NA        NA
#> 40         NA        NA        NA        NA        NA
#> 41         NA        NA        NA        NA        NA
#> 42         NA        NA        NA        NA        NA
#> 43         NA        NA        NA        NA        NA
#> 44         NA        NA        NA        NA        NA
#> 45         NA        NA        NA        NA        NA
#> 46         NA        NA        NA        NA        NA
#> 47         NA        NA        NA        NA        NA
#> 48         NA        NA        NA        NA        NA
#> 49         NA        NA        NA        NA        NA
#> 50         NA        NA        NA        NA        NA
#> 51         NA        NA        NA        NA        NA
#> 52         NA        NA        NA        NA        NA
#> 53         NA        NA        NA        NA        NA
#> 54         NA        NA        NA        NA        NA
#> 55         NA        NA        NA        NA        NA
#> 56         NA        NA        NA        NA        NA
#> 57         NA        NA        NA        NA        NA
#> 58         NA        NA        NA        NA        NA
#> 59         NA        NA        NA        NA        NA
#> 60         NA        NA        NA        NA        NA
#> 61         NA        NA        NA        NA        NA
#> 62         NA        NA        NA        NA        NA
#> 63         NA        NA        NA        NA        NA
#> 64         NA        NA        NA        NA        NA
#> 65         NA        NA        NA        NA        NA
#> 66         NA        NA        NA        NA        NA
#> 67         NA        NA        NA        NA        NA
#> 68         NA        NA        NA        NA        NA
#> 69         NA        NA        NA        NA        NA
#> 70         NA        NA        NA        NA        NA
#> 71         NA        NA        NA        NA        NA
#> 72         NA        NA        NA        NA        NA
#> 73         NA        NA        NA        NA        NA
#> 74         NA        NA        NA        NA        NA
#> 75         NA        NA        NA        NA        NA
#> 76         NA        NA        NA        NA        NA
#> 77         NA        NA        NA        NA        NA
#> 78         NA        NA        NA        NA        NA
#> 79         NA        NA        NA        NA        NA
#> 80         NA        NA        NA        NA        NA
#> 81         NA        NA        NA        NA        NA
#> 82         NA        NA        NA        NA        NA
#> 83         NA        NA        NA        NA        NA
#> 84         NA        NA        NA        NA        NA
#> 85         NA        NA        NA        NA        NA
#> 86         NA        NA        NA        NA        NA
#> 87         NA        NA        NA        NA        NA
#> 88         NA        NA        NA        NA        NA
#> 89         NA        NA        NA        NA        NA
#> 90         NA        NA        NA        NA        NA
#> 91         NA        NA        NA        NA        NA
#> 92         NA        NA        NA        NA        NA
#> 93         NA        NA        NA        NA        NA
#> 94         NA        NA        NA        NA        NA
#> 95         NA        NA        NA        NA        NA
#> 96         NA        NA        NA        NA        NA
#> 97         NA        NA        NA        NA        NA
#> 98         NA        NA        NA        NA        NA
#> 99         NA        NA        NA        NA        NA
#> 100        NA        NA        NA        NA        NA