Roll Up Mass Properties and Uncertainties

'rollup_mass_props()' rolls up mass properties in a data frame with (at least) these columns: id, mass, Cx, Cy, Cz, Ixx, Iyy, Izz, Ixy, Ixz, Iyz, POIconv, Ipoint, σ_mass, σ_Cx, σ_Cy, σ_Cz, σ_Ixx, σ_Iyy, σ_Izz, σ_Ixy, σ_Ixz, σ_Iyz.

Usage

rollup_mass_props_and_unc(
  tree,
  df,
  validate_df = validate_mass_props_and_unc_table,
  ...
)

Arguments

tree

An igraph directed graph that is a valid single-rooted in-tree with edges from child vertex to parent vertex.

df

A mass properties and uncertainties table

validate_df

A validator for the tree and table, default validate_mass_props_and_unc_table()

...

Other parameters passed rollup()

Value

The updated data frame

Examples

rollup_mass_props_and_unc(sawe_tree, sawe_table)
#>         id  mass       Cx         Cy          Cz      Ixx      Iyy      Izz
#> 1   Widget 57.83 121.2000  0.0400000 -0.16000000 7258.900  8607.02 10453.40
#> 2 2nd Part 16.80  70.9000 -0.9500000  0.46000000   65.070  1124.65  1078.82
#> 3 Combined 74.63 109.8769 -0.1828594 -0.02043146 7341.733 42673.75 44482.05
#>        Ixy       Ixz       Iyz  σ_mass    σ_Cx      σ_Cy       σ_Cz    σ_Ixx
#> 1  834.440 -1198.380 -1066.580 1.24160 0.27640 0.2085000 0.06690000 386.9233
#> 2   76.010   202.830    13.620 1.73080 0.62340 0.5173000 0.14050000  12.4687
#> 3 1558.714 -1401.534 -1060.951 2.13008 0.95821 0.1999847 0.06178402 387.4017
#>       σ_Iyy     σ_Izz     σ_Ixy    σ_Ixz    σ_Iyz Ipoint POIconv
#> 1  171.4792  414.5547 1440.5402 344.6237 124.6860  FALSE       +
#> 2  109.1324  108.5481   55.8879 212.1241  11.5408  FALSE       +
#> 3 2789.3133 2815.3260 1488.0948 418.6048 125.3175  FALSE       +