Set mass properties for a row in a data frame

set_mass_props() sets mass properties for a specified row in a data frame with an id column.

Usage

set_mass_props(df, id, v)

Arguments

df

A data frame

id

ID value of the desired row

v

#' A list with the following named elements:

mass mass (numeric)

center_mass center of mass (3-dimensional numeric)

inertia Inertia tensor (3x3 numeric matrix)

point Logical indicating point mass, i.e., negligible inertia

poi_conv Enumeration c("+", "-") indicating sign convention for products of inertia

Value

The updated data frame

Examples

df <- data.frame(id = c("C.1.2.2.3.1.2.3", "C.1.2.2.3.2.1.1"))
v <- get_mass_props(mp_table, "C.1.2.2.3.2.1.1")
v$poi_conv = "+"
df <- set_mass_props(df, "C.1.2.2.3.2.1.1", v)
get_mass_props(df, "C.1.2.2.3.2.1.1")
#> $mass
#> [1] 0.3413137
#> 
#> $center_mass
#>          x          y          z 
#>  34.274465 -44.982071   5.211377 
#> 
#> $inertia
#>               x             y             z
#> x  4.986894e-01 -4.436985e-02 -4.096337e-05
#> y -4.436985e-02  4.528114e-01 -4.510443e-06
#> z -4.096337e-05 -4.510443e-06  4.583186e-01
#> 
#> $point
#> [1] FALSE
#>