Update mass properties

update_mass_props() updates mass properties for a specified target row from specified source rows in a data frame with (at least) these columns: id, mass, Cx, Cy, Cz, Ixx, Iyy, Izz, Ixy, Ixz, Iyz, POIconv, Ipoint.

Usage

update_mass_props(df, target, sources, override = set_poi_conv_from_target)

Arguments

df

A data frame

target

ID of the target row

sources

IDs of the source rows

override

An override function, called as override(df, target, value)

Value

The updated data fram

Examples

leaves <- test_table[which(!is.na(test_table$mass)), "id"]
df <- update_mass_props(test_table, "A.1", leaves)